Updated to be Error Free
there was a glitch with the L and Line Block but this commit which is almost identical to the last is actually Error Free now.

boundaries are set up. rotation is checked when blocks are close to the sides so they dont rotate out of bounds. Bottom isnt being check for rotation though, have to do some testing on the bottom of the boundaries/play field.